Core: improved NGX_ALIGNMENT detection on some x86_64 platforms.
Previously, we detected the default pool alignment as `sizeof(unsigned)`,
expecting that it would result in a platform word size.  Certain 64-bit
platforms prove this assumption wrong by keeping the `long` type 32-bit,
which is fully compliant with the C standard.

Use a real pointer-sized type to set the default NGX_ALIGNMENT.

As we override the detection in `auto/os/conf` for all the machine types
except x86, and Unix-like x86 systems prefer 64-bit `long`, the impact
of this change should be limited to Win64.