Charset filter: improved validation of charset_map with utf-8.
It was possible to write outside of the buffer boundary used to keep
UTF-8 decoded values when parsing conversion table configuration.

As it happens before UTF-8 decoding, the fix is to check in advance
if character codes are of more than 3-byte sequence.  Note that this
is already enforced later by checking ngx_utf8_decode() decoded values
for 0xffff, which corresponds to the maximum value encoded as a valid
3-byte sequence, so the fix should not result in functional changes.

Found with AddressSanitizer.
Fixes GitHub issue #529.